Up to step 10
At this point we are neither finished with the front-end nor do we have
any working API endpoints. We could have focused on building just the
front-end in the first half of the tutorial and then the back-end in
the second half of the tutorial, or vice versa, but personally, I have
never built an app like that. I typically go back and forth between
back-end and front-end parts of the application during my development
flow.
We can’t display any characters until they are added to the database.
In order to add new characters to the database we need to build a UI
for it and implement an API endpoint. That’s exactly what we will do
next.
Step 11. Add Character Component
This component consists of a simple form with a text field, radio
buttons and a submit button. Success and error messages will be
displayed within help-block under the text field.